fix bignums and add sys.typeof

+++ source/lang/dext/repl.d
@@ -15,8 +15,7 @@ import std.functional;
 /// vm callback that sets the locals defined into the root base 
 LocalCallback exportLocalsToBaseCallback(Function func)
 {
-    LocalCallback ret = (ref size_t index, ref size_t depth,
-            ref Dynamic[] stack, ref Dynamic[] locals) {
+    LocalCallback ret = (ref uint index, ref uint depth, ref Dynamic[] stack, ref Dynamic[] locals) {
         foreach (i, ref v; locals[0 .. func.stab.byPlace.length])
         {
             rootBase ~= Pair(func.stab.byPlace[i], v);
@@ -26,7 +25,7 @@ LocalCallback exportLocalsToBaseCallback(Function func)
 }
 
 /// vm callback that prints the top of the stack for the end of the repl
-void printTop(ref size_t index, ref size_t depth, ref Dynamic[] stack, ref Dynamic[] locals)
+void printTop(ref uint index, ref uint depth, ref Dynamic[] stack, ref Dynamic[] locals)
 {
     writeln(stack[depth]);
 }
